Fix public key URL
Github recently started serving raw content from a different domain, so this cURL command no longer works. The old domain redirects to the new domain, but because cURL was not invoked with the -L flag, it does not follow the redirect.

The URL has been updated to the new address.